+++ src/ciencias/ii/búsqueda/en/grafos/launcher/Launcher.java
@@ -1,0 +1,12 @@
+package ciencias.ii.búsqueda.en.grafos.launcher;
+
+import ciencias.ii.búsqueda.en.grafos.control.Controlador;
+
+public class Launcher {
+
+	public static void main(String[] args) {
+		Controlador c = new Controlador();
+		c.iniciar();
+	}
+
+}
